+++ .nuget/Changelog.md
@@ -1,1 +1,10 @@
-Revit 2023 support
+New JoinGeometryUtils extensions
+
+Geometry extensions
+
+- New line.SetCoordinateX extension
+- New line.SetCoordinateY extension
+- New line.SetCoordinateZ extension
+- New arc.SetCoordinateX extension
+- New arc.SetCoordinateY extension
+- New arc.SetCoordinateZ extension