+++ Changelog.md
@@ -1,3 +1,16 @@
+# Release 2023.1.1
+
+New JoinGeometryUtils extensions
+
+Geometry extensions
+
+- New line.SetCoordinateX extension
+- New line.SetCoordinateY extension
+- New line.SetCoordinateZ extension
+- New arc.SetCoordinateX extension
+- New arc.SetCoordinateY extension
+- New arc.SetCoordinateZ extension
+
 # Release 2023.1.0
 
 - Revit 2023 support